// src/app/models/user.ts
export class User {
    id: number = 0;
    username: string = "";
    password: string = "";
    firstname: string = "";
    lastname: string = "";
    type: string = "";
    gender: string = "";
    address: string = "";
    contact: string = "";
    email: string = "";
    profilePicture: File = new File([], "");
    creditCard: string = "";
    status: string = "";
};