// src/app/models/registration.ts
export class Registration {
    id: number = 0;
    username: string = "";
    password: string = "";
    firstname: string = "";
    lastname: string = "";
    type: string = "";
    gender: string = "";
    address: string = "";
    contact: string = "";
    email: string = "";
    profilePicture: string = "";
    profilePictureName: string | undefined;
    creditCard: string = "";
    status: string = "";
}